refactor: ChatService 역할 분리 및 컴포넌트화를 통한 구조 개선
- 비즈니스 로직이 집중된 ChatService의 책임을 세부 컴포넌트로 분리
- **ChatReader**: 채팅 히스토리 조회 및 데이터 읽기 전담 컴포넌트 생성
- **ChatWriter**: 채팅 메시지, 메타데이터, 토큰 사용량 저장 로직을 캡슐화한 컴포넌트 생성
- **ChatStreamHandler**: AI 응답 파싱 및 SSE(Server-Sent-Events) 변환 로직 분리
- **ChatMapper**: Entity와 DTO 간의 변환 로직 분리
- **ChatStreamContext**: 스트리밍 중 상태 관리를 위한 컨텍스트 객체 추가
- ChatService는 흐름 제어(Orchestration) 역할만 수행하도록 로직 경량화
